added the feature to select detection from stamps and lightcurve
